refactor: inline wavReader into fileLoader
