+++ modules/fileLoader.js
@@ -1,8 +1,29 @@
 // modules/fileLoader.js
 
 import { extractGuanoMetadata, parseGuanoMetadata } from './guanoReader.js';
-import { getWavSampleRate } from "./wavReader.js";
 import { addFilesToList, getFileList, getCurrentIndex, setCurrentIndex, removeFilesByName, setFileMetadata } from './fileState.js';
+
+export async function getWavSampleRate(file) {
+  if (!file) return 256000;
+  const buffer = await file.arrayBuffer();
+  const view = new DataView(buffer);
+  let pos = 12;
+  while (pos < view.byteLength - 8) {
+    const chunkId = String.fromCharCode(
+      view.getUint8(pos),
+      view.getUint8(pos + 1),
+      view.getUint8(pos + 2),
+      view.getUint8(pos + 3)
+    );
+    const chunkSize = view.getUint32(pos + 4, true);
+    if (chunkId === 'fmt ') {
+      return view.getUint32(pos + 12, true);
+    }
+    pos += 8 + chunkSize;
+    if (chunkSize % 2 === 1) pos += 1; // word alignment
+  }
+  return 256000;
+}
 
 let lastObjectUrl = null;
